+++ AzureCP/AzureCP.cs
@@ -849,8 +849,7 @@ namespace azurecp
             this.Lock_Config.EnterReadLock();
             try
             {
-                int maxCount = 30;  // SharePoint sets maxCount to 30 in method FillSearch
-                OperationContext currentContext = new OperationContext(CurrentConfiguration, OperationType.Search, ProcessedClaimTypesList, resolveInput, null, context, entityTypes, null, maxCount);
+                OperationContext currentContext = new OperationContext(CurrentConfiguration, OperationType.Search, ProcessedClaimTypesList, resolveInput, null, context, entityTypes, null, CurrentConfiguration.MaxSearchResultsCount);
                 List<PickerEntity> entities = SearchOrValidate(currentContext);
                 FillEntities(currentContext, ref entities);
                 if (entities == null || entities.Count == 0) return;