+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 * Improve tests
 * Use Azure DevOps to build AzureCP
 * Cache result returned by FileVersionInfo.GetVersionInfo() to avoid potential hangs
+* Add property AzureCPConfig.MaxSearchResultsCount to set max number of results returned to SharePoint during a search
 * Remove reference on Microsoft.IdentityModel.Clients.ActiveDirectory.Platform.dll
 * Update NuGet package Microsoft.Graph to v1.12
 * Update NuGet package Microsoft.IdentityModel.Clients.ActiveDirectory to v4.4.2